Refactor backend and frontend to display announcements and interact with database
In this commit, we made several changes to both the backend and frontend of the Zalzal App. These changes include:

1. Modified backend files:
   - `app.js`
   - `productController.js`
   - `userController.js`
   - `userModel.js`
   - `productRoute.js`
   - `userRoute.js`
   - `server.js`

2. Deleted `productModel.js` from the backend.

3. Modified frontend files:
   - `package.json`
   - `public/index.html`
   - `src/actions/userAction.js`
   - `src/components/Home/Map.jsx`
   - `src/components/Home/Menu.jsx`
   - `store.js`

These changes were made to transition from creating announcements to displaying them and handling interactions with the database. The modifications encompass various aspects of the application, including routing, data retrieval, and user actions related to announcements.